% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj.R
\docType{data}
\name{adj}
\alias{adj}
\title{Adjacent relationship}
\format{
\describe{
\item{g}{group id}
\item{adj}{adjacent group id}
}
}
\usage{
adj
}
\description{
The \code{adj} is adjacency of groups in \code{crimetko} with 2078 rows and 2 variables.
}
\keyword{datasets}
