% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GFLglm.R
\name{GFLglm}
\alias{GFLglm}
\title{Generalized fused Lasso for generalized linear models}
\usage{
GFLglm(
  y,
  group,
  D,
  family,
  link = NULL,
  alpha = NULL,
  offset = FALSE,
  adaptive = TRUE,
  as = NULL,
  nlam = 100,
  maxit = 500,
  lam.cand = NULL,
  dsp.nb = 2,
  tol = 1e-05,
  progress = FALSE
)
}
\arguments{
\item{y}{a vector of a response variable}

\item{group}{a vector of groups}

\item{D}{a list of adjacent relationship among groups}

\item{family}{the error distribution; one of "Gaussian", "Binomial", "Poisson",
"Negative.Binomial" (or "NB"), "Gamma", and "Inverse.Gaussian" (or "IG")}

\item{link}{the link function; the default is canonical link;
this can take "log" for "Gamma", "NB", and "IG"}

\item{alpha}{a positive value adjusting the strength of the penalty of
generalized information criterion, e.g., 2 corresponds to AIC;
the default is \code{log(length(y))} which corresponds to BIC}

\item{offset}{a vector of offset}

\item{adaptive}{if \code{TRUE}, penalty weights are used}

\item{as}{a vector of the numbers of trials for only \code{family=Binomial}}

\item{nlam}{the number of tuning parameter candidates}

\item{maxit}{an upper bound for iteration}

\item{lam.cand}{if "uniform" is given, tuning parameter candidates are defined
by uniformly dividing}

\item{dsp.nb}{for only "NB", this decides the estimation method of dispersion parameter;
1: by Pearson estimator; 2: using Pearson estimator (default; see, Table 8.4 of
Hilbe, 2011);
3: Newton's method; 4: only once update by update equation of Newton's method}

\item{tol}{a tolerance for convergence}

\item{progress}{if \code{TRUE}, progress is displayed}
}
\value{
a list having following elements:
\item{fitted.values}{fitted values}

\item{beta}{estimates of group parameters}

\item{eta}{linear predictors}

\item{phi}{dispersion parameter}

\item{cluster}{cluster of groups}

\item{offset}{offset}

\item{weight}{penalty weights}

\item{results}{results for each tuning parameter candidate which have the
following columns: \code{lambda} (tuning parameter), \code{MSC} (model selection criterion value),
\code{NLL} (nagative log-likelihood function value), \code{DF} (degrees of freedom; the number of clusters),
\code{dispersion} (dispersion parameter), \code{chisq} (Pearson's chi squared statistic),
\code{convergence} (algorithm converged (1) or not (0))}

\item{all.beta}{estimates of group parameters for each tuning parameter candidate}

\item{summary}{results under the optimal tuning parameter, which has
\code{idx} (index of the optimal tuning parameter), \code{lambda} (the optimal tuning parameter),
\code{dispersion} (dispersion parameter), \code{df} (the number of clusters),
\code{family}, \code{link}, \code{convergence}, \code{time} (total runtime)}
}
\description{
\code{GFLglm} This function provides generalized fused Lasso estimator for
generalized linear models via coordinate descent algorithm (v0.1.0)
}
\examples{
#GFLglm(y, group, D, family)
}
\references{
Hilbe, J. M. (2011). \emph{Negative Binomial Regression 2nd Edition}.
Cambridge University Press. Cambridge.

Ohishi, M. (2024).
Generalized fused Lasso for grouped data in generalized linear models.
\emph{Hiroshima Statistical Research Group Technical Report}, TR-No. 24-02, Hiroshima University.
}
